pyproject.toml
[tool.poetry]
name = "option-diffusion"
version = "0.1.0"
description = "Conditional Diffusion Model for Option Markets Hedging"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "option_diffusion"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
pytorch-lightning = "^2.0.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
boto3 = "^1.28.0"
mlflow = "^2.8.0"
hydra-core = "^1.3.0"
psycopg2-binary = "^2.9.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
dvc = {extras = ["s3"], version = "^3.0.0"}
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.1.0"
pre-commit = "^3.5.0"
ipython = "^8.0.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["option_diffusion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
